movie-playlist-site/backend/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
numpy==1.21.2
pandas==1.3.3
lightfm==1.16
scikit-learn==0.24.2
python-dotenv==0.19.0
supabase==0.0.2 